[test] add tests for moved constant expressions
